requirements.txt
asgiref==3.5.2
Django==4.1.1
djangorestframework==3.13.1
psycopg2-binary==2.9.3
python-dateutil==2.8.2
python-dotenv==0.21.0
pytz==2022.2.1
six==1.16.0
sqlparse==0.4.2